% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sheet_data.R
\name{clean_sheet_data}
\alias{clean_sheet_data}
\title{Clean ASBQ & PSQI Google Sheet data}
\usage{
clean_sheet_data(urlstring)
}
\arguments{
\item{urlstring}{Input url string of Google Sheet data. Ensure that sharing is enabled on the GSheet is enabled}
}
\value{
A csv file called 'asbq_psqi_df.csv' in an output folder called 'output'
}
\description{
Clean ASBQ & PSQI Google Sheet data
}
\examples{

clean_sheet_data("https://docs.google.com/spreadsheets/d/1cnb_5DUQsbee96lL_5MVtf_nI8XmJqKmYQKFP9_INJY/edit?usp=sharing")

}
