% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dashboard.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Generate an interactive Shiny dashboard app with ASBQ & PSQI Google Sheet data k-means clustering}
\usage{
create_dashboard(urlstring)
}
\arguments{
\item{urlstring}{Input url string of Google Sheet data. Ensure that sharing is enabled on the GSheet is enabled}
}
\value{
A Shiny interactive app
}
\description{
Generate an interactive Shiny dashboard app with ASBQ & PSQI Google Sheet data k-means clustering
}
\examples{

create_dashboard("https://docs.google.com/spreadsheets/d/1cnb_5DUQsbee96lL_5MVtf_nI8XmJqKmYQKFP9_INJY/edit?usp=sharing")

}
