% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_sheet_data.R
\name{clean_sheet_data}
\alias{clean_sheet_data}
\title{Clean ASBQ & PSQI Google Sheet data}
\usage{
clean_sheet_data(
  urlstring,
  morning_endtime = "06:00:00",
  morning_correction_start = "06:01:00",
  morning_correction_end = "11:59:00"
)
}
\arguments{
\item{urlstring}{Input url string of Google Sheet data. Ensure that sharing is enabled on the GSheet is enabled}

\item{morning_endtime}{Input time string to indicate the morning cutoff time. Default is '06:00:00'}

\item{morning_correction_start}{Input time string to indicate the start time for corrections to occur to correct potential participant error in bedtime inputs. E.g. if participant indicates a bedtime of 7 am as a bedtime, this will be corrected to a 'pm' value. Default is '06:01:00'}

\item{morning_correction_end}{Input time string to indicate the end time for corrections to occur to correct potential participant error in bedtime inputs. E.g. if participant indicates a bedtime of 7 am as a bedtime, this will be corrected to a 'pm' value. Default is '11:59:00'}
}
\value{
A csv file called 'asbq_psqi_df.csv' in an output folder called 'output'
}
\description{
Clean ASBQ & PSQI Google Sheet data
}
\examples{

clean_sheet_data("https://docs.google.com/spreadsheets/d/1cnb_5DUQsbee96lL_5MVtf_nI8XmJqKmYQKFP9_INJY/edit?usp=sharing")

}
